Small refactoring

+++ app/src/main/java/com/unica/so2/enotesrecorder/Helper/JsonHelper.java
@@ -2,15 +2,17 @@ package com.unica.so2.enotesrecorder.Helper;
 
 
 import com.unica.so2.enotesrecorder.Model.Content;
+import com.unica.so2.enotesrecorder.Model.Note;
 
 import org.json.JSONObject;
 
-import java.io.FileOutputStream;
-import java.io.ObjectOutputStream;
-
 public class JsonHelper {
 
-        /*serialization: from Content to Json string*/
+       /** This method implements the serialization from the Content, passed as parameter,
+        *  to a Json String
+        *
+        * @return String, throws an exception otherwise
+        * */
         public static String serializeContent(Content content) {
 
             String jsonString = "";
@@ -28,7 +30,11 @@ public class JsonHelper {
             return jsonString;
         }
 
-        // deserialization: from Json string to Content
+    /** This method implements the deserialization from a String, passed as parameter,
+     *  to a Content
+     *
+     * @return Content, throws an exception otherwise
+     * */
         public static Content deserializeContent(String jsonString) {
             Content content = new Content();
 
@@ -43,8 +49,13 @@ public class JsonHelper {
 
             return content;
         }
-    
-        /*serialization: from Note to Json string*/
+
+
+    /** This method implements the serialization from the Note, passed as parameter,
+     *  to a Json String
+     *
+     * @return String of the jsonObject created, throws an exception otherwise
+     * */
         public static String serializeNote(Note note) {
     
             String jsonString = "";
@@ -64,9 +75,14 @@ public class JsonHelper {
     
             return jsonString;
         }
-    
-        // deserialization: from Json string to Note
-        public static Note deserializeNote(String jsonString) {
+
+
+    /** This method implements the deserialization from the json String, passed as parameter,
+     *  to a Note
+     *
+     * @return Note , throws an exception otherwise
+     * */
+    public static Note deserializeNote(String jsonString) {
             Note note = new Note();
     
             try {
@@ -75,7 +91,7 @@ public class JsonHelper {
                 note.setTitle(json.getString("title"));
                 note.setContent(deserializeContent(json.getString("content")));
                 note.setLastEdit(GenericHelper.stringToDate(json.getString("lastEdit")));
-                note.setRating(json.getDouble("rating"));
+                note.setRating((float)json.getDouble("rating"));
             }
             catch (Exception e) {
                 e.printStackTrace();
